started with templ and serving pages
