+++ internal/server/response.go
@@ -1,4 +1,4 @@
-package api
+package server
 
 import (
 	"encoding/json"
