+++ internal/templates/imprint.templ
@@ -1,0 +1,7 @@
+package templates
+
+templ Imprint(imprintText string) {
+    @layout() {
+        <div class="text-center">{ imprintText }</div>
+    }
+}
