+++ static/static.go
@@ -1,0 +1,6 @@
+package static
+
+import "embed"
+
+//go:embed *
+var Dir embed.FS
